XPath - select - hiddenElement

+++ src/Utility/MyFunc.java
@@ -3,9 +3,9 @@ package Utility;
 import java.util.Scanner;
 
 public class MyFunc {
-    public static void Bekle(int saniye){
+    public static void Bekle(double saniye){
         try {
-            Thread.sleep(saniye*1000);
+            Thread.sleep((long) (saniye*1000));
         } catch (InterruptedException e) {
             throw new RuntimeException(e);
         }
